f/MUWM-4844: made section and linked-section two different components
