+++ myuw/static/vue/components/home/inst_course_summary/meeting.vue
@@ -153,6 +153,10 @@
 <script>
 export default {
   props: {
+    mobileOnly: {
+      type: Boolean,
+      default: false,
+    },
     section: {
       type: Object,
       required: true,
