+++ myuw/static/vue/home.js
@@ -1,5 +1,4 @@
 import {Vue, vueConf} from './base.js';
-import utils from './mixins/utils.js';
 
 // layout componenets
 import Boilerplate from './components/_templates/boilerplate/boilerplate.vue';
